requirements.txt
aiohttp
pytest
beautifulsoup4
requests
sqlalchemy
pyyaml
APScheduler
psycopg2-binary